pyproject.toml
[tool.poetry]
name = "githubapiapp"
version = "0.1.0"
description = ""
authors = ["Kieran Pritchard <kieran.pritchard@ons.gov.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
datetime = "^5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
